requirements.txt
networkx
matplotlibgit 